# doc/swagger/la-candela-backend-rpc.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "La Candela Backend RPC",
    "description": "La Candela Backend RPC",
    "version": "1.0.0",
    "contact": {
      "name": "Emeka Okoro",
      "email": "okoroemeka056@gmail.com"
    }
  },
  "tags": [
    {
      "name": "LaCandelaBackendRPC"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/create_course": {
      "post": {
        "operationId": "LaCandelaBackendRPC_CreateCourse",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbCreateCourseResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbCreateCourseRequest"
            }
          }
        ],
        "tags": [
          "LaCandelaBackendRPC"
        ]
      }
    },
    "/v1/create_user": {
      "post": {
        "summary": "Create a new user",
        "description": "Create a new user",
        "operationId": "LaCandelaBackendRPC_CreateUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbCreateUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbCreateUserRequest"
            }
          }
        ],
        "tags": [
          "LaCandelaBackendRPC"
        ]
      }
    },
    "/v1/login": {
      "post": {
        "summary": "login user",
        "description": "Login existing user",
        "operationId": "LaCandelaBackendRPC_LoginUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbLoginUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbLoginUserRequest"
            }
          }
        ],
        "tags": [
          "LaCandelaBackendRPC"
        ]
      }
    }
  },
  "definitions": {
    "pbCourse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "courseTitle": {
          "type": "string"
        },
        "courseLevel": {
          "type": "string"
        },
        "startDate": {
          "type": "string",
          "format": "date-time"
        },
        "endDate": {
          "type": "string",
          "format": "date-time"
        },
        "startTime": {
          "type": "string",
          "format": "date-time"
        },
        "endTime": {
          "type": "string",
          "format": "date-time"
        },
        "price": {
          "type": "string",
          "format": "int64"
        },
        "locationId": {
          "type": "string"
        },
        "minCapacity": {
          "type": "integer",
          "format": "int32"
        },
        "open": {
          "type": "boolean"
        },
        "priceHikeId": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "createdBy": {
          "type": "string"
        },
        "courseDescription": {
          "type": "string"
        }
      }
    },
    "pbCreateCourseRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "courseTitle": {
          "type": "string"
        },
        "courseLevel": {
          "type": "string"
        },
        "startDate": {
          "type": "string",
          "format": "date-time"
        },
        "endDate": {
          "type": "string",
          "format": "date-time"
        },
        "startTime": {
          "type": "string",
          "format": "date-time"
        },
        "endTime": {
          "type": "string",
          "format": "date-time"
        },
        "price": {
          "type": "string",
          "format": "int64"
        },
        "locationId": {
          "type": "string"
        },
        "minCapacity": {
          "type": "integer",
          "format": "int32"
        },
        "open": {
          "type": "boolean"
        },
        "priceHikeId": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "createdBy": {
          "type": "string"
        },
        "courseDescription": {
          "type": "string"
        }
      }
    },
    "pbCreateCourseResponse": {
      "type": "object",
      "properties": {
        "course": {
          "$ref": "#/definitions/pbCourse"
        }
      }
    },
    "pbCreateUserRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        }
      }
    },
    "pbCreateUserResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/pbUser"
        }
      }
    },
    "pbLoginUserRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "pbLoginUserResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/pbUser"
        }
      }
    },
    "pbUser": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "passwordChangedAt": {
          "type": "string",
          "format": "date-time"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "isEmailVerified": {
          "type": "boolean"
        },
        "id": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}